requirements.txt
dedalus-mcp>=0.7.0
dedalus-labs>=0.2.0
pyyaml>=6.0
jinja2>=3.1
click>=8.1
httpx>=0.28
jsonschema>=4.20
pydantic>=2.12
python-dotenv>=1.0
